requirements.txt
langchain 
langchain-community
langchain-openai
langchain-google-genai
python-dotenv
pydantic